# openapi.yaml
components:
  headers:
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        minimum: 1
        type: integer
  parameters:
...
    ChallengeId:
      description: The challenge, as carried on its challenge line.
      in: path
      name: challengeId
      required: true
      schema:
        type: string
...
  responses:
...
    Banned:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BannedError"
      description: The bot's owner is banned; after the ban lifts, the owner must rotate the token.
    BotUnauthorized:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
      description: Missing, unknown, or rotated bot token.
...
    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
      description: The target does not exist or is not the caller's to act on.
...
    Paused:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PausedError"
      description: "The site is paused: nothing new starts, while open streams and live games run on.
        Retry after 60 s."
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
...
  schemas:
...
    BannedError:
      properties:
        code:
          enum:
            - banned
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    Challenge:
      description: firstPlayer names who plays turn (openingPlies + 1) / 2, the first turn after the
        opening. A random firstPlayer is drawn at acceptance.
      properties:
        challengeId:
          type: string
        challenger:
          $ref: "#/components/schemas/Player"
        destUser:
          $ref: "#/components/schemas/Player"
        firstPlayer:
          enum:
            - challenger
            - challenged
            - random
          type: string
        openingPlies:
          $ref: "#/components/schemas/OpeningPlies"
        status:
          enum:
            - created
            - accepted
            - declined
            - canceled
            - expired
          type: string
        timeControl:
          $ref: "#/components/schemas/TimeControl"
      required:
        - challengeId
        - challenger
        - destUser
        - timeControl
        - openingPlies
        - firstPlayer
        - status
      type: object
    ChallengeAcceptError:
      properties:
        code:
          enum:
            - bad_request
            - bot_busy
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    ChallengeCreateError:
      properties:
        code:
          enum:
            - bad_request
            - not_open
            - clock_not_accepted
            - bot_busy
            - inbox_full
            - daily_pair_cap
            - daily_bot_cap
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    ChallengeForbiddenError:
      properties:
        code:
          enum:
            - banned
            - own_bot
            - delisted
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    NotFoundError:
      properties:
        code:
          enum:
            - not_found
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    Ok:
      properties:
        ok:
          enum:
            - true
          type: boolean
      required:
        - ok
      type: object
    OpeningPlies:
      description: Stones the server places before either player acts, the origin at (0, 0) included. A
        ply is one stone; turn 0 is the origin, and turn t >= 1 is plies 2t-1 and 2t. Each later ply
        lands in order on an empty cell drawn uniformly from those within hex distance 2 of the
        origin, for the player who owns that ply. The whole opening is redrawn while any 6
        consecutive cells on an axis hold 4 or more stones of one player and none of the other. The
        draw uses a CSPRNG when the game is created, for a challenge at acceptance.
      enum:
        - 1
        - 3
        - 5
        - 7
        - 9
      type: integer
    PausedError:
      properties:
        code:
          enum:
            - paused
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    Player:
      properties:
        name:
          type: string
        provisional:
          $ref: "#/components/schemas/Provisional"
        rating:
          $ref: "#/components/schemas/Rating"
      required:
        - name
        - rating
        - provisional
      type: object
    Provisional:
      description: True while the rating deviation is above 75.
      type: boolean
    Rating:
      description: The Glicko-2 rating, rounded to a whole point.
      type: integer
...
    TimeControl:
      description: Durations are in milliseconds.
      oneOf:
        - properties:
            mode:
              enum:
                - unlimited
              type: string
          required:
            - mode
          type: object
        - properties:
            mode:
              enum:
                - turn
              type: string
            turnTimeMs:
              minimum: 5000
              type: integer
          required:
            - mode
            - turnTimeMs
          type: object
        - properties:
            incrementMs:
              minimum: 0
              type: integer
            mainTimeMs:
              minimum: 60000
              type: integer
            mode:
              enum:
                - match
              type: string
          required:
            - mode
            - mainTimeMs
            - incrementMs
          type: object
    UnauthorizedError:
      properties:
        code:
          enum:
            - unauthorized
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: opaque
      description: A bot token (hxo_...), minted and rotated by the owner on the website.
      scheme: bearer
      type: http
    gameToken:
      bearerFormat: opaque
      description: A game token (hgs_...) from a gameStart line.
      scheme: bearer
      type: http
    sessionCookie:
      description: An HttpOnly session cookie, set by the Discord OAuth callback or the guest route.
      in: cookie
      name: hexo_arena_session
      type: apiKey
info:
  title: HeXO Arena
  version: 0.19.0
openapi: 3.0.3
paths:
...
  /api/bot/challenge/{challengeId}/accept:
    post:
      description: Only the challenged bot may accept. The game starts at once, and both bots receive gameStart.
      operationId: acceptChallenge
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
          description: Accepted.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeAcceptError"
          description: The challenged bot is at its game cap (bot_busy); the challenge stays pending.
        "401":
          $ref: "#/components/responses/BotUnauthorized"
        "403":
          $ref: "#/components/responses/Banned"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Paused"
      security:
        - bearerAuth: []
      summary: Accept a challenge as the challenged bot.
      tags:
        - Challenge
  /api/bot/challenge/{challengeId}/cancel:
    post:
      description: Only the challenger may cancel; the target receives challengeCanceled.
      operationId: cancelChallenge
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
          description: Canceled.
        "401":
          $ref: "#/components/responses/BotUnauthorized"
        "403":
          $ref: "#/components/responses/Banned"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - bearerAuth: []
      summary: Cancel a challenge the caller issued.
      tags:
        - Challenge
  /api/bot/challenge/{challengeId}/decline:
    post:
      description: Only the challenged bot may decline; the challenger receives challengeDeclined.
      operationId: declineChallenge
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
          description: Declined.
        "401":
          $ref: "#/components/responses/BotUnauthorized"
        "403":
          $ref: "#/components/responses/Banned"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - bearerAuth: []
      summary: Decline a challenge as the challenged bot.
      tags:
        - Challenge
  /api/bot/challenge/{name}:
    post:
      description: The target must hold its stream open with open=1 and accept the clock. A bot plays at
        most 4 games at once. Per UTC day, a bot plays at most 100 bot-vs-bot games, and a pair 20.
        The target holds at most 10 pending challenges. A challenge expires after 60 s. Resending a
        requestId answers 200 with the stored challenge.
      operationId: createChallenge
      parameters:
        - description: The challenged bot's name.
          in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                firstPlayer:
                  default: random
                  enum:
                    - challenger
                    - challenged
                    - random
                  type: string
                openingPlies:
                  allOf:
                    - $ref: "#/components/schemas/OpeningPlies"
                    - default: 5
                requestId:
                  maxLength: 128
                  minLength: 1
                  type: string
                timeControl:
                  $ref: "#/components/schemas/TimeControl"
              required:
                - timeControl
                - requestId
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
          description: The stored challenge for a known requestId, with the status it reached.
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
          description: The challenge, pending in the target's inbox.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeCreateError"
          description: Validation failed (bad_request), the target is not open (not_open) or excludes the
            clock (clock_not_accepted), a side is at its game cap (bot_busy), the target's inbox is
            full (inbox_full), or the pair (daily_pair_cap) or a bot (daily_bot_cap) is at its daily
            cap.
        "401":
          $ref: "#/components/responses/BotUnauthorized"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeForbiddenError"
          description: The challenger's owner owns the target (own_bot), either bot is delisted (delisted), or
            the challenger's owner is banned (banned).
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Paused"
      security:
        - bearerAuth: []
      summary: Challenge another bot.
      tags:
        - Challenge
...